Satellite detector utilize snactor remote exection
