+++ src/actors/satellite-detector/_actor.yaml
@@ -1,22 +1,12 @@
-inputs:
-- name: host
-  type:
-    name: BaseTypeString
-- name: user
-  type:
-    name: BaseTypeString
-
 outputs:
 - name: app_satellite_detection 
   type:
     name: AppSatelliteDetectionResult
+    version: 1.0
 
 description: |
   Detect whether there is satellite installed on the host 
 
-  Inputs:
-    host: remote host address
-    user: remote user name
   Outputs:
     app_satellite_detection - a structure containing information about detected satellite
