+++ src/actors/satellite-detector/satellite_detector_actor.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+import requests
 import sys
 
 import json
@@ -6,51 +7,108 @@ import json
 from subprocess import Popen, PIPE
 import shlex
 
-# Defaults inputs
-inputs = {
-    "user": "root",
-    "host": "localhost"
-}
 
-# Output fields
-output_fields = {"satellite_detected": False,
-                 "satellite_version_major": None,
-                 "satellite_version_minor": None,
-                 "satellite_version_bugfix": None,
-                 "satellite_services": []}
-
-if not sys.stdin.isatty():
-    input_data = sys.stdin.read()
-    if input_data:
-        loaded_inputs = json.loads(input_data)
-else:
-    loaded_inputs = {}
-
-# Load inputs
-for key in inputs.keys():
-    if key in loaded_inputs.keys():
-        inputs[key] = loaded_inputs[key][0].get("value", inputs[key])
+class DetectionFailed(Exception):
+    pass
 
 
-mode = "local" if inputs["host"] in ("localhost", "127.0.0.1") else "ssh"
-cmd = "ansible -m satellite_detector -M . -i {host}, -c {mode} -u {user} {host}".format(host=inputs["host"],
-                                                                                        user=inputs["user"],
-                                                                                        mode=mode)
-
-ansible = Popen(shlex.split(cmd), stdout=PIPE)
-ansible_stdout, _ = ansible.communicate()
-
-if not ansible.returncode:
-    ansible_json = json.loads(ansible_stdout.split("=>")[1])
-    output = {}
-
-    if not ansible_json.get("satellite_detected", False):
-        output["satellite_detected"] = False
-    else:
-        for key in output_fields.keys():
-            output[key] = ansible_json.get(key, output_fields[key])
-
-    print(json.dumps({"app_satellite_detection": [output]}))
+def extract_version(version_string, result):
+    version_sep = version_string.split(".")
+    result["satellite_version_major"] = version_sep[0]
+    result["satellite_version_minor"] = version_sep[1]
+    result["satellite_version_bugfix"] = ".".join(version_sep[2:])
 
 
-sys.exit(ansible.returncode)
+def execute_cmd(cmd, shell=False):
+    if not shell:
+        cmd = shlex.split(cmd)
+
+    process = Popen(cmd, stdout=PIPE, stderr=PIPE, shell=shell)
+    out, err = process.communicate()
+
+    return (process.returncode, out, err)
+
+
+def detect_satellite5():
+    # TODO: get this information from rpm actor (input rpm)
+    verif_code, verif_out, _ = execute_cmd("rpm -q satellite-schema --qf %{VERSION}")
+
+    if not verif_code:
+        result = dict(
+            satellite_detected=True,
+        )
+
+        # Version
+        extract_version(verif_out, result)
+
+        # Services
+        services_code, serv_out, _ = execute_cmd(
+            cmd='rhn-satellite list | grep -e \":on\" -e \":off\" | awk \'{print $1}\'',
+            shell=True)
+
+        if services_code or serv_out is None:
+            raise DetectionFailed("Satellite services could not be detected")
+
+        result["satellite_services"] = serv_out.split("\n")
+
+        # Optional services
+        # TODO: get this information from chkconfig actor (input chkconfig)
+        services_code, _, _ = execute_cmd("chkconfig --list dhcpd | grep on", shell=True)
+        if not services_code:
+            result["satellite_services"].append("dhcp")
+
+        services_code, _, _ = execute_cmd("chkconfig --list --type xinetd tftp | grep on", shell=True)
+        if not services_code:
+            result["satellite_services"].append("tftp")
+
+        return result
+
+    return None
+
+
+def detect_satellite6():
+    # TODO: get this information from rpm actor (input rpm)
+    verif_code, verif_out, _ = execute_cmd("rpm -q satellite --qf %{VERSION}")
+
+    if not verif_code:
+        result = dict(
+            satellite_detected=True,
+        )
+
+        extract_version(verif_out, result)
+
+        services_code, serv_out, _ = execute_cmd(
+            cmd="echo -n $(katello-service list | sed -e 's/^\\([^[:space:]]\\+\\).*/\\1/g' -e 's/\\.service$//')",
+            shell=True)
+
+        if services_code or serv_out is None:
+            raise DetectionFailed("Satellite services could not be detected")
+
+        result["satellite_services"] = serv_out.split(" ")
+        features = []
+
+        try:
+            features = list(requests.get("https://localhost:9090/features", verify=False).json())
+        except Exception:
+            raise DetectionFailed("Capsule features could not be detected")
+
+        result["satellite_services"] = list(set(result["satellite_services"] + features))
+        return result
+
+    return None
+
+
+def detect_satellite():
+    sat5_detection = detect_satellite5()
+    if sat5_detection is not None:
+        return sat5_detection
+
+    sat6_detection = detect_satellite6()
+    if sat6_detection is not None:
+        return sat6_detection
+
+    return {"satellite_detected": False}
+
+
+if __name__ == '__main__':
+    print(json.dumps({"app_satellite_detection": [detect_satellite()]}))
