+++ src/schemas/type_app_satellite_detection_result.py
